Add get Methods in GameObject and modify name velocity into speed because is ambiguous.

+++ ModernProject/Player.h
@@ -8,12 +8,15 @@
 class Player:public GameObject
 {
 public:
-	Player( std::string_view name,std::pair<uint16_t,uint16_t>position,float velocity);
+	Player( std::string_view name,std::pair<uint16_t,uint16_t>position,double velocity);
 	void Shoot(uint8_t direction);
 	// Method which a players shooting another players;
 	void ResetPosition();
 	// Method which player respawn in startPoint;
 	void Display() override;
+
+
+
 private:
 	std::string_view m_name;
 	std::uint16_t m_health;
